implementations/uim-chatbot/requirements.txt
# Core Framework
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.10.3
pydantic-ai==0.0.14

# HTTP Client
httpx==0.28.1

# Logging
loguru==0.7.3

# Environment Variables
python-dotenv==1.0.1

# XML Parsing (for arXiv API responses)
xmltodict==0.14.2
